requirements.txt
aiohttp~=3.10.10
aiogram~=3.15.0
PyYAML~=6.0.2
